Fix server to use Channel instead of stdin/stdout
