+++ denops/denops/mod.ts
@@ -1,4 +1,3 @@
-export * from "./cli.ts";
 export * from "./context.ts";
 export * from "./service.ts";
 export * as host from "./host/mod.ts";
